Added batteryCheck() to Rules.
Today: added chargers and charger messages. Added charger (an charger vector) to "alice_mail", the alice struct "mail", and as a field called "chargers" in model. Added appropraite subscriber to hub. Created a new exeutable, and made Hawk_Sim as its own class, which only has charger stuff setup so far, but will be easy to add to.